[raphael-cli] Move commands to their own files
