+++ raphael-cli/src/commands/mod.rs
@@ -1,0 +1,2 @@
+pub mod search;
+pub mod solve;
