+++ raphael-cli/src/main.rs
@@ -1,7 +1,6 @@
 use clap::{Parser, Subcommand};
-use game_data::{get_game_settings, get_item_name, CrafterStats, Locale, RECIPES};
-use simulator::SimulationState;
-use solvers::MacroSolver;
+
+mod commands;
 
 #[derive(Parser, Debug)]
 #[command(
@@ -78,23 +77,7 @@ fn main() {
 
     match args.command {
         Commands::Search { pattern } => {
-            let output_separator = std::env::var("OFS").unwrap_or(" ".to_string());
-            let matches = game_data::find_recipes(&pattern, Locale::EN);
-            if matches.is_empty() {
-                println!("No matches found");
-                return;
-            }
-
-            for recipe_idx in matches {
-                let recipe = &RECIPES[recipe_idx];
-                let name = get_item_name(recipe.item_id, false, Locale::EN);
-                println!(
-                    "{item_id}{separator}{name}",
-                    item_id = recipe.item_id,
-                    separator = output_separator,
-                    name = name
-                );
-            }
+            commands::search::execute(pattern);
         }
         Commands::Solve {
             item_id,
@@ -109,12 +92,8 @@ fn main() {
             backload_progress,
             unsound,
         } => {
-            let recipe = RECIPES
-                .iter()
-                .find(|r| r.item_id == item_id)
-                .expect("Recipe not found");
-
-            let crafter_stats = CrafterStats {
+            commands::solve::execute(
+                item_id,
                 craftsmanship,
                 control,
                 cp,
@@ -122,37 +101,10 @@ fn main() {
                 manipulation,
                 heart_and_soul,
                 quick_innovation,
-            };
-
-            let settings = get_game_settings(*recipe, crafter_stats, None, None, adversarial);
-            let state = SimulationState::new(&settings);
-
-            let mut solver = MacroSolver::new(
-                settings,
+                adversarial,
                 backload_progress,
                 unsound,
-                Box::new(|_| {}),
-                Box::new(|_| {}),
             );
-            let actions = solver.solve(state).expect("Failed to solve");
-
-            let final_state = SimulationState::from_macro(&settings, &actions).unwrap();
-            let quality = final_state.quality;
-            let steps = actions.len();
-            let duration: i16 = actions.iter().map(|action| action.time_cost()).sum();
-
-            println!("Item ID: {}", recipe.item_id);
-            println!("Quality: {}/{}", quality, settings.max_quality);
-            println!(
-                "Progress: {}/{}",
-                final_state.progress, settings.max_progress
-            );
-            println!("Steps: {}", steps);
-            println!("Duration: {} seconds", duration);
-            println!("\nActions:");
-            for action in actions {
-                println!("{:?}", action);
-            }
         }
     }
 }
